renamed to bonusEffects
