+++ web/src/game/engine.ts
@@ -7,7 +7,7 @@ import { processAttacks } from './engine/combat'
 import { BASE_MAX_MANA, BLOOD_POOL_FADE_MS, MANA_REGEN_MS, OPPONENT_INTERVAL_MS, PLAYER_SPAWN_X, SPAWN_GROW_MS } from './engine/constants'
 import { genericBossAI, getBossAIDef } from './engine/boss'
 import { tickBossTrait } from './engine/bossTraits'
-import { getManaBonus, getManaSpeedMult } from './engine/getManaBonus'
+import { getManaBonus, getManaSpeedMult } from './engine/bonusEffects'
 import { uid, shuffle, spawnUnit } from './engine/helpers'
 import { opponentAI } from './engine/opponentAI'
 import { triggerBattleEvent, BATTLE_EVENT_BASE_MS } from './engine/battleEvents'
